# More efficient decoding of booleans