Tested with Scala 3 and JDK-21 on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                                   (size)   Mode  Cnt        Score       Error  Units
ArrayBufferOfBooleansReading.zioSchemaJson     128  thrpt    5  1048786.538 ±  6630.269  ops/s
ArrayOfBooleansReading.zioJson                 128  thrpt    5  1884086.300 ± 10715.550  ops/s
ArrayOfBooleansReading.zioSchemaJson           128  thrpt    5  1226574.624 ± 17789.201  ops/s
ArraySeqOfBooleansReading.zioJson              128  thrpt    5  1687694.071 ± 39218.973  ops/s
ArraySeqOfBooleansReading.zioSchemaJson        128  thrpt    5  1131214.620 ± 12011.729  ops/s
ListOfBooleansReading.zioJson                  128  thrpt    5  1183228.227 ± 15484.292  ops/s
ListOfBooleansReading.zioSchemaJson            128  thrpt    5   801988.755 ±  8861.996  ops/s
VectorOfBooleansReading.zioJson                128  thrpt    5  1445047.569 ± 33342.106  ops/s
VectorOfBooleansReading.zioSchemaJson          128  thrpt    5   828646.802 ±  8922.681  ops/s
```

#### After
```txt
Benchmark                                   (size)   Mode  Cnt        Score        Error  Units
ArrayBufferOfBooleansReading.zioSchemaJson     128  thrpt    5  1097902.384 ±  10064.691  ops/s
ArrayOfBooleansReading.zioJson                 128  thrpt    5  1892103.362 ±  23129.647  ops/s
ArrayOfBooleansReading.zioSchemaJson           128  thrpt    5  1229650.481 ±  15709.329  ops/s
ArraySeqOfBooleansReading.zioJson              128  thrpt    5  1812104.561 ± 134985.997  ops/s
ArraySeqOfBooleansReading.zioSchemaJson        128  thrpt    5  1200243.600 ±  39887.079  ops/s
ListOfBooleansReading.zioJson                  128  thrpt    5  1199166.956 ±  76466.575  ops/s
ListOfBooleansReading.zioSchemaJson            128  thrpt    5   829455.128 ±  22945.978  ops/s
VectorOfBooleansReading.zioJson                128  thrpt    5  1480336.259 ±  20425.230  ops/s
VectorOfBooleansReading.zioSchemaJson          128  thrpt    5   863184.010 ±  31147.887  ops/s
```